requirements.txt
# ===== Backend =====
fastapi==0.127.0
uvicorn[standard]==0.40.0
pydantic==2.12.5
python-dotenv==1.2.1
requests==2.32.5

# ===== LangChain Stack (COMPATIBLE) =====
langchain==1.1.0
langchain-core==1.2.7
langchain-community==0.4.1
langchain-openai==1.1.7

# ===== OpenAI =====
openai==2.14.0

# ===== Vector DB =====
chromadb==1.3.5

# ===== ML =====
numpy==1.26.4
scipy==1.16.3
torch==2.9.1
transformers==4.36.2

# ===== Docs / Utils =====
python-docx==1.1.0
pymupdf==1.26.7
python-Levenshtein==0.27.3


python-multipart==0.0.21